previously, when a reminder is set to less than today, it will show the date of the reminder(in the past), but when it's today, it will show the time. It now shows time when it is less than today aswell.
Reminders can now be postponed.

The reminder listview will now be updated after you close the reminder popup form. This is because reminders will now be updated after you close to popup

Fixed a bug where the date of an newly made (weekly)reminder would be an week in the future even if the selected time is larger than the current time
i.e. an user creates a reminder every thursday at 11:00am(selected), and at the time of creation it is thursday, 10:50 pm(current).
The first reminder should then be in 10 minutes, not in a week and 10 minutes