V1.5.02
Refactored some code that didn't belong in BLIO and moved it to an method in DLReminder instead

Refactored the weekly type of reminder. removed it since there's now the weekdays option where the user can select multiple week days, so only one, too. Which has the same functionality as the previous "weekly" reminder