V1.5.04
Added a new method SaveAndCloseDataBase() to DLReminder. This simply saves and disposes the database, but also refreshes the local list. everytime when the database needs saving, instead of calling RemindMeDbEntities.SaveChanges, we call this method. Then, the local cache list also gets updated.

fixed a small bug where if you pressed the "x" button on a reminder preview popup, it would say that the reminder doesn't exist(because the id is -1, and that doesn't exist in the database)