V1.5.1 - Big update 2 New features!
1)RemindMe now has a custom "scroll" message that contains a custom message and uses an animation to scroll up from the bottom right corner of RemindMe.

2)Monthly reminders can now contain multiple days each month. Example: you can create a reminder for each 1st, 10th and 20th of the month.

Because monthly reminders can now contain multiple days, a LOT of code had to be rewritten.
The date datetimepicker shows the earliest date from the selected monthly days. example: monthly days:1,5,10. If it's the 3th of the month, it picks the 5th.

The version label of RemindMe now gets the version from the assembly, so the label doesnt have to get updated manually

More unit tests