V1.6.0 - Big update! Removed the settings icon from RemindMe's contextmenustrip, as it no longer launches a new form.
Fixed an issue where when you caused the scrolling popup to come up for 2 different reasons before the scrolling popup has fully gone down, you would still see the old text of the first reason it came up.
The size was adjusted for the text of the new reason, but you would still see the old reason's text because
it added 2 controls into the panel instead of clearing it first. It clears the controls when it fully goes down, this is why this issue happened.

Also made it so that if the scrolling popup is visible for a reason, it won't start scrolling from the bottom again if you did something wrong that triggers
the popup again with the same reason.

-------------------------------------------------------
The big part of this update:
RemindMe has been partly rewritten to make use of a Data Access Layer and a Business Logic Layer. This is purely for OOP purposes, the user of this program will not notice this.

The Data Access Layer contains all logic for database-side methods. The Business Logic Layer interacts with this layer, and only the Business Logic Layer. No other class may use the Data Access Layer. The User Interface classes my call the Business Logic Layer, which then will call the Data Access Layer.

Note: in order for other classes to use the Reminder class for methods like
-MakePopup(Reminder rem)
-FillControlsForEdit(Reminder rem)
The project RemindMeReWrite has to add the reference Database(project with the db model). Which, could make use of RemindMeDbEntities and edit the database like that.
I've decided to do it like this since i think it's necessary for this project(which contains all the UI classes aswell)
to make use of the database classes Reminder,Settings and Songs.