V1.6.3 .remindme file extension support
Associated the file type .remindme to RemindMe
Reminders can now be exported & imported from settings. When pressing the export button, it creates an .remindme file in the selected folder.
This .remindme file can be imported again by going to the settings and pressing the import button

Also: When you double-click a .remindme file, it will launch the RemindMeImporter form.
If RemindMe wasn't running yet, it launches RemindMe and you'd have to double-click the .remindme again
From this form, you can import reminders

Fixed an issue with creating a shortcut of RemindMe

When a reminder pops up with a sound that cannot be played(the file is not at the saved destination anymore) it is removed from the reminder, so that the user
won't get the same popup multiple times whenever that reminder pops up again