V1.6.31 Improved .remindme functionality
.remindme files are now written to a serialized file. this is a solution where you don't need a lot of code, so a lot of existing code could be removed.
This way, you can just easily call SerializeRemindersToFile(list of reminders, path to the file) and that will write all the reminders in the list, serialized to a .remindme file.
Then, to read the file, you just simply call DeserializeRemindersFromFile(path to the file) and that method returns a List<Reminder>.

2 methods with not a lot of code in them and easy to use!
Another plus, if you right click the .remindme file and open it with notepad, you will mostly see weird null's and icons

When importing reminders, the soundfilepath property of the reminder will now be removed if the file does not exist on the target computer.