Fixed a small issue where resetting the Popup size/font by pressing the reset button would result in the numeric control having the value of the previous value So, if the title font size was 18, then you save it to 17, then you press reset to defaults, the numeric control would show 18. It was correctly saved as 14 in the database though.
RemindMe's custom error form now looks a little bit nicer and automatically enlarges on larger errors.

Fixed an issue when you try to export reminders to a folder that you do not have rights to save to. It now shows a message that says RemindMe is not authorized to save to that path without administrator mode