V1.7.21
You will now be notified of reminders that will happen in the next hour. If RemindMe is running, and while checking reminders it detects that one or multiple
reminders will happen within 60 minutes(not less, 50,40, etc) it will show the RemindMe scrolling form that appears on the right bottom part of your screen
with the name(s) of the reminder(s). This option is toggleable in the settings.