V1.7.24 You can now add/subtract days from a date by pressing the ... button next to the datetime picker. It then launches the prompt described below.
RemindMe now has a custom prompt where users can fill in a number or a string. The prompt will then return those values (RemindMePrompt.cs)

Reminders that would be deleted if they wouldn't be repeated are now being archived. For example if there is a reminder set for a certain date, and the reminder
pops up, the reminder would have been deleted afterwards. It now gets archived instead.