V1.7.29
You can now add minutes to the current time when you are creating an reminder(from the ... button)
This can be used whenever you want to create a new reminder in for example 40 minutes

Fixed an issue with settings custom sizes for RemindMe's Popup in the settings. The preview feature didn't work as intended. It now works again