V2.1.4
Fixed an issue with importing reminders from the double-click .remindme option option. It would import them but the remindme list wouldn't refresh to show the changes.

Fixed an memory issue. When you would edit a lot of reminders, the edit-reminder form wouldn't get disposed, so this would build up. It now disposes it correctly.